+++ apps/nextjs-website/src/app/layout.tsx
@@ -1,11 +1,11 @@
 import React from 'react';
 import type { Metadata } from 'next';
 import { PreHeaderProps } from '@pagopa/pagopa-editorial-components/dist/components/PreHeader';
-import { PreHeaderClient } from '@/lib/components/PreHeaderClient';
-import { getPreHeaderData } from '@/lib/API/preHeaderAPI';
+import { PreHeaderClient } from '@/components/PreHeaderClient';
+import { getPreHeaderData } from '@/lib/api/preHeaderAPI';
 import { HeaderProps } from '@pagopa/pagopa-editorial-components/dist/components/Header/Header';
-import { HeaderClient } from '@/lib/components/HeaderClient';
-import { getHeaderData } from '@/lib/API/HeaderAPI';
+import { HeaderClient } from '@/components/HeaderClient';
+import { getHeaderData } from '@/lib/api/HeaderAPI';
 
 export const metadata: Metadata = {
   title: 'Page',
