--- apps/nextjs-website/src/lib/Reusable/z-declaration.ts
@@ -1,7 +1,0 @@
-import { z } from 'zod';
-
-export const ThemeSchema = z.enum(['light', 'dark']);
-
-export const CTAButtonVariant = z.enum(['text', 'outlined', 'contained']);
-
-export const CTAButtonColor = z.enum(['inherit', 'primary', 'secondary', 'success', 'error', 'info', 'warning']);
